Moves Endpoint class into its own file.
