+++ packages/subiquity_client/lib/subiquity_client.dart
@@ -1,5 +1,6 @@
 library subiquity_client;
 
 export 'src/client.dart' hide log;
+export 'src/endpoint.dart';
 export 'src/status_monitor.dart' hide log;
 export 'src/types.dart';
