+++ packages/subiquity_client/lib/subiquity_server.dart
@@ -1,3 +1,4 @@
 library subiquity_server;
 
+export 'src/endpoint.dart';
 export 'src/server.dart' hide log;
